+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - import an existing `did:web` document into a `did:webvh` via `DidWebImporter`
+- add support for parallel `did:web` publishing via `DidWebPublisher`
 
 
 ## [0.2.1] - 2026-05-19